+++ programs/name-registry/src/constants.rs
@@ -1,0 +1,6 @@
+const NAME_LEN: usize = 20;
+const LINK_LEN: usize = 160;
+const BIO_LEN: usize = 160;
+const TITLE_LEN: usize = 80;
+
+pub const REGISTRATION_FEE_LAMPORTS: u64 = 100_000;
