+++ programs/name-registry/src/instructions/mod.rs
@@ -1,8 +1,10 @@
+pub mod initialize;
 pub mod register_name;
 pub mod reserve_name;
 pub mod set_metadata;
 pub mod transfer_name;
 
+pub use initialize::*;
 pub use register_name::*;
 pub use reserve_name::*;
 pub use set_metadata::*;
