+++ programs/name-registry/src/lib.rs
@@ -11,11 +11,7 @@ pub mod name_registry {
     use super::*;
 
     pub fn initialize(ctx: Context<Initialize>) -> Result<()> {
-        let config = &mut ctx.accounts.config;
-        config.admin = *ctx.accounts.user.key;
-
-        msg!("Initialized with id: {:?}", ctx.program_id);
-        Ok(())
+        initialize::initialize(ctx)
     }
 
     pub fn register_name(ctx: Context<RegisterName>) -> Result<()> {
@@ -33,26 +29,4 @@ pub mod name_registry {
     pub fn reserve_name(ctx: Context<ReserveName>) -> Result<()> {
         reserve_name::reserve_name(ctx)
     }
-}
-
-#[account]
-pub struct Config {
-    pub admin: Pubkey,
-}
-
-#[derive(Accounts)]
-pub struct Initialize<'info> {
-    #[account(
-      init,
-      payer = payer,
-      space = 8 + 32,
-      seeds = [b"config"],
-      bump
-    )]
-    pub config: Account<'info, Config>,
-    #[account(mut)]
-    pub payer: Signer<'info>,
-    #[account(mut)]
-    pub user: Signer<'info>, // initial admin
-    pub system_program: Program<'info, System>,
 }
